requirements.txt
click
numpy
Pillow
tensorflow-gpu>=1.0.0